Fix for build success.

+++ backend/remill/include/remill/Arch/Runtime/Intrinsics.h
@@ -41,7 +41,7 @@ extern "C" {
 
 [[gnu::used]] extern void __remill_write_memory_64(RuntimeManager *, addr_t, uint64_t);
 
-[[gnu::used]] extern void __remill_write_memory_128(RuntimeManager *, addr_t, uint64_t);
+[[gnu::used]] extern void __remill_write_memory_128(RuntimeManager *, addr_t, uint128_t);
 
 [[gnu::used, gnu::pure]] extern float32_t __remill_read_memory_f32(RuntimeManager *, addr_t);
